// market-flipper-agent/src/transaction-actor/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartBucket, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  API_ROUTER: ServiceStub<import('../api-router/index.js').default>;
  DECISION_ACTOR: ActorNamespace<import('../decision-actor/index.js').DecisionActor>;
  FLIPPER_INVENTORY: SmartBucket;
  FLIPPER_LISTINGS: SmartBucket;
  FLIPPER_TRANSACTIONS: SmartBucket;
  HISTORICAL_PRICES: SmartBucket;
  INVENTORY_ACTOR: ActorNamespace<import('../inventory-actor/index.js').InventoryActor>;
  logger: Logger;
  SCRAPER_SERVICE: ServiceStub<import('../scraper-service/index.js').default>;
  tracer: Tracer;
  TRANSACTION_ACTOR: ActorNamespace<import('./index.js').TransactionActor>;
  VALUATION_SERVICE: ServiceStub<import('../valuation-service/index.js').default>;
}